refactor: 나뉘어진 media query 사용 방식 통일
