+++ components/Button/styled.ts
@@ -2,7 +2,7 @@ import styled, { css } from "styled-components";
 import { BrandColorTypes, theme } from "../../styles/Theme";
 import React from "react";
 import { darken, lighten, transparentize } from "polished";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 
 export interface ButtonStyle
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
