+++ components/Skeleton/ManageCardSkeleton/styled.ts
@@ -1,5 +1,5 @@
 import styled, { keyframes } from "styled-components";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 const loading = keyframes`
  0% {
       background-color: rgba(165, 165, 165, 0.1);
