+++ components/Skeleton/ProjectCardSkeleton/styled.ts
@@ -1,5 +1,5 @@
 import styled, { keyframes } from "styled-components";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 import { theme } from "../../../styles/Theme";
 
 const loading = keyframes`
