+++ components/Skeleton/RecruitCardSkeleton/styled.ts
@@ -1,5 +1,5 @@
 import { theme } from "@/styles/Theme";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 import styled, { keyframes } from "styled-components";
 
 const loading = keyframes`
