+++ components/Tag/TagBox/styled.ts
@@ -1,4 +1,4 @@
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 import styled, { css } from "styled-components";
 
 export const TagContainer = styled.div`
