+++ components/pages/landing/second/styled.ts
@@ -1,5 +1,5 @@
 import styled, { keyframes } from "styled-components";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 
 const slideAnimationMain = keyframes`
   0% {
