+++ components/pages/landing/styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 export const SectionContainer = styled.section`
   height: calc(100vh - 75px);
   display: flex;
