+++ components/pages/mypage/TabBoards/styled.ts
@@ -1,4 +1,4 @@
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 import styled from "styled-components";
 
 export const BoardWrapper = styled.div`
