+++ components/pages/mypageEdit/Introduction/styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 
 export const ProfileWrapper = styled.div`
   width: 100%;
