+++ components/pages/project/ProjectLIst/styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { media } from "@/styles/mediatest";
+import { media } from "@/styles/media";
 export const ProjectListWrapper = styled.main`
   display: grid;
   justify-items: center;
