+++ components/pages/recruits/RecruitList/styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { mediaQuery } from "@/styles/Media";
+import { media } from "@/styles/media";
 
 export const Wrapper = styled.div`
   width: 100%;
@@ -9,10 +9,10 @@ export const Wrapper = styled.div`
   row-gap: 1rem;
   column-gap: 1rem;
 
-  ${mediaQuery("mobile")`
-  display: flex;
-  flex-direction: column;
-`}
+  ${media.mobile} {
+    display: flex;
+    flex-direction: column;
+  }
 `;
 
 export const NoListWrapper = styled.div`
