+++ styles/media.ts
@@ -1,10 +1,12 @@
-const sizes = {
+export const sizes = {
   mobile: 768,
   laptop: 1024,
   desktop: 1280,
 };
+
 const customMediaQuery = (maxWidth: number) =>
   `@media (max-width: ${maxWidth}px)`;
+
 export const media = {
   custom: customMediaQuery,
   mobile: customMediaQuery(sizes.mobile),
